refactor: mostly nuke common.
