--- rgdps/constants/xor.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-
-class XorKeys:
-    LEVEL_PASSWORD = b"26364"
-    MESSAGE = b"14251"
-    QUESTS = b"19847"
-    CHESTS = b"59182"
