+++ rgdps/utilities/colour.py
